+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v0.6.6]
+
+### Added
+- Add support for array operators in CQL2 (a_equals, a_contains, a_contained_by, a_overlaps).
+- Add check in loader to make sure that pypgstac and pgstac versions match before loading data [#123](https://github.com/stac-utils/pgstac/issues/123)
+
 ## [v0.6.5]
 
 ### Fixed
@@ -200,7 +206,8 @@ _TODO_
 
 - Fixed issue with pypgstac loads which caused some writes to fail ([#18](https://github.com/stac-utils/pgstac/pull/18))
 
-[unreleased]: https://github.com/stac-utils/pgstac/compare/v0.6.5...HEAD
+[unreleased]: https://github.com/stac-utils/pgstac/compare/v0.6.6...HEAD
+[v0.6.6]: https://github.com//stac-utils/pgstac/compare/v0.6.5...v0.6.6
 [v0.6.5]: https://github.com//stac-utils/pgstac/compare/v0.6.4...v0.6.5
 [v0.6.4]: https://github.com//stac-utils/pgstac/compare/v0.6.3...v0.6.4
 [v0.6.3]: https://github.com//stac-utils/pgstac/compare/v0.6.2...v0.6.3